refactor(docsite): make inline code run a list of linkifiers
Replaces the bespoke `astryx docs` command parser with a linkifier step on
the inline-code renderer: a list of (code) => href | null functions, tried in
order. Adding a new kind of cross-reference is now writing a function and
appending it to LINKIFIERS.

The parser it replaces handled runner prefixes and binaries no doc actually
uses. The whole corpus is two forms — `astryx docs <topic>` and one with an
`npx` prefix — so the single rule is a regex, and the resolver drops from 83
lines to a third of that with no behaviour change.
